Issue 1770 bugfix large inputs calendar 3.3.1 - closes #1770 (#2074)
* limit max value for Unix-Time to 26.1.29349 00:00:00
correct tests for valid DateTime Dates

* check and correct possible overflow
add testst

* check and correct overflow
add tests

* add tests

* refactor file names

+++ test/tools/science_and_technology/date_and_time/calendar_week/logic/calendar_week_test.dart
@@ -88,7 +88,7 @@ void main() {
     for (var elem in _inputsToExpected) {
       test('date: ${elem['date']}, iso: ${elem['iso']}', () {
         var _actual = calendarWeek(elem['date'] as DateTime, iso: elem['iso'] as bool);
-        expect(_actual, (elem['expectedOutput'] as int));
+        expect(_actual, elem['expectedOutput']);
       });
     }
   });
